Functionality Change
The university Degreeworks website does not provide a PDF of a
student’s courses and grades. Instead, it opens up a pop-up window
which can be saved as a XML file - saving the file does not work (seems
to be security related).

Converted PDF option to quick insert option, where GPA is calculated
after every entry and course name is ignored.