Revised option 2 - GPACalculator
Fooling around with different ways of entering information. I liked not
being prompted 3 times for different chunks of information. Writing it
all down at once makes for a smoother input process.

Functional but not pretty. Contemplating on adding an option to save
entered information to a plain text file, working with file IO.

Plans to also improve on prompts and input process.